demo/backend/requirements.txt
# Core dependencies
streamlit>=1.30.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-chroma>=0.1.0

# Vector store and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Translation and multilingual support
googletrans>=4.0.0
deep-translator>=1.11.0
langdetect>=1.0.9

# Document processing
pypdf>=3.0.0
python-docx>=0.8.0
python-pptx>=0.6.0
beautifulsoup4>=4.12.0
unstructured>=0.10.0

# API and HTTP
openai>=1.0.0
requests>=2.31.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.7.0

# Japanese text processing
janome>=0.4.0
mecab-python3>=1.0.0 